+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [#214](https://github.com/wallix/awless/issues/214): `AWS_PROFILE` env variable now loaded in `awless` in addition to the deprecated `AWS_DEFAULT_PROFILE` thanks to @alewando
 - Better completion for `attach mfadevice` and `attach user` commands
+- [#219](https://github.com/wallix/awless/issues/219): Validate access key and secret key before writing into `~/.aws/credentials` file
 
 ### Fixes